+++ src/drivers/warmers/LocalWarmer.php
@@ -10,7 +10,6 @@ use craft\helpers\App;
 use craft\web\UrlManager;
 use craft\web\UrlRule;
 use putyourlightson\blitz\Blitz;
-use putyourlightson\blitz\events\RefreshCacheEvent;
 use putyourlightson\blitz\helpers\CacheWarmerHelper;
 use putyourlightson\blitz\models\SiteUriModel;
 
@@ -41,10 +40,7 @@ class LocalWarmer extends BaseCacheWarmer
      */
     public function warmUris(array $siteUris, int $delay = null, callable $setProgressHandler = null)
     {
-        $event = new RefreshCacheEvent(['siteUris' => $siteUris]);
-        $this->trigger(self::EVENT_BEFORE_WARM_CACHE, $event);
-
-        if (!$event->isValid) {
+        if (!$this->beforeWarmCache($siteUris)) {
             return;
         }
 
@@ -55,9 +51,7 @@ class LocalWarmer extends BaseCacheWarmer
             CacheWarmerHelper::addWarmerJob($siteUris, 'warmUrisWithProgress', $delay);
         }
 
-        if ($this->hasEventHandlers(self::EVENT_AFTER_WARM_CACHE)) {
-            $this->trigger(self::EVENT_AFTER_WARM_CACHE, $event);
-        }
+        $this->afterWarmCache($siteUris);
     }
 
     /**